updated readme link

+++ src/vulnerabilities/reentrancy/classic/solution/SealedOasis.sol
@@ -10,8 +10,7 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
 
 contract SealedOasis is ReentrancyGuard {
-    
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
     function deposit() public payable {
         require(msg.value > 0, "Deposit must be greater than 0");
@@ -20,12 +19,12 @@ contract SealedOasis is ReentrancyGuard {
 
     // CEI applied: balance updated (Effect) before ETH is sent (Interaction)
     // nonReentrant modifier adds a second layer — any reentrant call hits a locked door
-    function withdraw(uint _amount) public nonReentrant {
+    function withdraw(uint256 _amount) public nonReentrant {
         require(_amount <= balances[msg.sender], "Insufficient balance"); // Check
 
         balances[msg.sender] -= _amount; // Effect
 
-        (bool success, ) = msg.sender.call{value: _amount}(""); // Interaction
+        (bool success,) = msg.sender.call{value: _amount}(""); // Interaction
         require(success, "Transfer failed.");
     }
 }
